requirements.txt
--extra-index-url https://pypi.tuna.tsinghua.edu.cn/simple
--find-links https://download.pytorch.org/whl/torch_stable.html
Flask==2.3.2
Flask-CORS==4.0.0
opencv-python==4.8.0.74
numpy==1.24.3
scikit-learn==1.0.2
onnxruntime-gpu==1.15.1
mxnet==1.7.0.post2
torch==2.0.1+cu118
torchvision==0.15.2+cu118
insightface==0.7.3
Cython==0.29.32
ImageHash==4.3.1
Pillow==10.0.0
waitress==2.0.0
faiss-gpu==1.7.4
